Fix minor bug causing collision reverting to not trigger
I was able to reliably reproduce this with the torch. I'm not sure why it was so easy to reproduce with the torch. To reproduce, walk to the right of the torch and position your y so it is just barely overlapping with the torch's hitbox. Walk left. This causes you to walk through the torches hitbox. A collision is detected but the x axis is not being properly backed out. I think this may be due to the way the hitTestRectangle is set up and will have to check that out in more detail.

This change does make it slightly less precise to walk next to an object and have it automatically move you to the right spot which I view as a positive upside to this.