Cap all object movement to playable area
This may not be a long term thing or it may be tweaked as time goes on. I can see the use-case for NPCs and items to move off screen / out of the playable area. Maybe add a property on the InteractableObject that can toggle this behavior (or define the boundary).