Replace collision side detection logic with different approach
I ended up trying a few different approaches myself but all of them were flawed in some way. This one that I found on SO works well with the existing logic which means it's probably the least flawed. I have not yet found a bug with it so I'm sticking with it for now.