Add support for predicates on event callbacks
A predicate will make it easier to specify a callback that is a "once" when it should only fire in certain instances. For example, a collision callback that only runs once and only when a player triggered the collision. In this case the predicate will check if the player was the cause of this collision and, if so, trigger the event and remove the callback.